+++ src/ru/henchan/src/eu/kanade/tachiyomi/extension/ru/henchan/Filters.kt
@@ -1,0 +1,221 @@
+package eu.kanade.tachiyomi.extension.ru.henchan
+
+import android.annotation.SuppressLint
+import eu.kanade.tachiyomi.source.model.Filter
+
+class Genre(
+    val id: String,
+    @SuppressLint("DefaultLocale") name: String = id.replace('_', ' ').capitalize(),
+) : Filter.TriState(name)
+
+class GenreList(genres: List<Genre>) : Filter.Group<Genre>("Тэги", genres)
+
+class OrderBy :
+    UriPartFilter(
+        "Сортировка",
+        arrayOf("Дата", "Популярность", "Алфавит"),
+        arrayOf("&n=dateasc" to "", "&n=favasc" to "&n=favdesc", "&n=abcdesc" to "&n=abcasc"),
+        arrayOf(
+            "manga/new&n=dateasc" to "manga/new",
+            "manga/new&n=favasc" to "mostfavorites&sort=manga",
+            "manga/new&n=abcdesc" to "manga/new&n=abcasc",
+        ),
+    )
+
+open class UriPartFilter(
+    displayName: String,
+    sortNames: Array<String>,
+    val withGenres: Array<Pair<String, String>>,
+    val withoutGenres: Array<Pair<String, String>>,
+) : Filter.Sort(displayName, sortNames, Selection(1, false)) {
+    fun toUriPartWithGenres() = if (state!!.ascending) withGenres[state!!.index].first else withGenres[state!!.index].second
+
+    fun toUriPartWithoutGenres() = if (state!!.ascending) withoutGenres[state!!.index].first else withoutGenres[state!!.index].second
+}
+
+internal fun getGenreList() = listOf(
+    Genre("3D"),
+    Genre("action"),
+    Genre("ahegao"),
+    Genre("bdsm"),
+    Genre("corruption"),
+    Genre("foot_fetish"),
+    Genre("footfuck"),
+    Genre("gender_bender"),
+    Genre("live"),
+    Genre("lolcon"),
+    Genre("megane"),
+    Genre("mind_break"),
+    Genre("monstergirl"),
+    Genre("netorare"),
+    Genre("netori"),
+    Genre("nipple_penetration"),
+    Genre("oyakodon"),
+    Genre("paizuri_(titsfuck)"),
+    Genre("rpg"),
+    Genre("scat"),
+    Genre("shemale"),
+    Genre("shimaidon"),
+    Genre("shooter"),
+    Genre("simulation"),
+    Genre("skinsuit"),
+    Genre("tomboy"),
+    Genre("tomgirl"),
+    Genre("x-ray"),
+    Genre("алкоголь"),
+    Genre("анал"),
+    Genre("андроид"),
+    Genre("анилингус"),
+    Genre("анимация"),
+    Genre("аркада"),
+    Genre("арт"),
+    Genre("бабушка"),
+    Genre("без_текста"),
+    Genre("без_трусиков"),
+    Genre("без_цензуры"),
+    Genre("беременность"),
+    Genre("бикини"),
+    Genre("близнецы"),
+    Genre("боди-арт"),
+    Genre("больница"),
+    Genre("большая_грудь"),
+    Genre("большие_попки"),
+    Genre("бондаж"),
+    Genre("буккаке"),
+    Genre("в_ванной"),
+    Genre("в_общественном_месте"),
+    Genre("в_первый_раз"),
+    Genre("в_цвете"),
+    Genre("в_школе"),
+    Genre("вампиры"),
+    Genre("веб"),
+    Genre("вебкам"),
+    Genre("вибратор"),
+    Genre("визуальная_новелла"),
+    Genre("внучка"),
+    Genre("волосатые_женщины"),
+    Genre("гаремник"),
+    Genre("гг_девушка"),
+    Genre("гг_парень"),
+    Genre("гипноз"),
+    Genre("глубокий_минет"),
+    Genre("горячий_источник"),
+    Genre("грудастая_лоли"),
+    Genre("групповой_секс"),
+    Genre("гяру_и_гангуро"),
+    Genre("двойное_проникновение"),
+    Genre("девочки_волшебницы"),
+    Genre("девушка_туалет"),
+    Genre("демоны"),
+    Genre("дилдо"),
+    Genre("дочь"),
+    Genre("драма"),
+    Genre("дыра_в_стене"),
+    Genre("жестокость"),
+    Genre("за_деньги"),
+    Genre("зомби"),
+    Genre("зрелые_женщины"),
+    Genre("измена"),
+    Genre("изнасилование"),
+    Genre("инопланетяне"),
+    Genre("инцест"),
+    Genre("исполнение_желаний"),
+    Genre("камера"),
+    Genre("квест"),
+    Genre("кимоно"),
+    Genre("колготки"),
+    Genre("комиксы"),
+    Genre("косплей"),
+    Genre("кремпай"),
+    Genre("кудере"),
+    Genre("кузина"),
+    Genre("куннилингус"),
+    Genre("купальники"),
+    Genre("латекс_и_кожа"),
+    Genre("магия"),
+    Genre("маленькая_грудь"),
+    Genre("мастурбация"),
+    Genre("мать"),
+    Genre("мейдочки"),
+    Genre("мерзкий_дядька"),
+    Genre("минет"),
+    Genre("много_девушек"),
+    Genre("молоко"),
+    Genre("монашки"),
+    Genre("монстры"),
+    Genre("мочеиспускание"),
+    Genre("мужская_озвучка"),
+    Genre("мужчина_крепкого_телосложения"),
+    Genre("мускулистые_женщины"),
+    Genre("на_природе"),
+    Genre("наблюдение"),
+    Genre("непрямой_инцест"),
+    Genre("новелла"),
+    Genre("обмен_партнерами"),
+    Genre("обмен_телами"),
+    Genre("обычный_секс"),
+    Genre("огромная_грудь"),
+    Genre("огромный_член"),
+    Genre("оплодотворение"),
+    Genre("остановка_времени"),
+    Genre("парень_пассив"),
+    Genre("переодевание"),
+    Genre("песочница"),
+    Genre("племянница"),
+    Genre("пляж"),
+    Genre("подглядывание"),
+    Genre("подчинение"),
+    Genre("похищение"),
+    Genre("презерватив"),
+    Genre("принуждение"),
+    Genre("прозрачная_одежда"),
+    Genre("проникновение_в_матку"),
+    Genre("психические_отклонения"),
+    Genre("публично"),
+    Genre("рабыни"),
+    Genre("романтика"),
+    Genre("сверхъестественное"),
+    Genre("секс_игрушки"),
+    Genre("сестра"),
+    Genre("сетакон"),
+    Genre("скрытный_секс"),
+    Genre("спортивная_форма"),
+    Genre("спящие"),
+    Genre("страпон"),
+    Genre("суккубы"),
+    Genre("темнокожие"),
+    Genre("тентакли"),
+    Genre("толстушки"),
+    Genre("трап"),
+    Genre("тётя"),
+    Genre("умеренная_жестокость"),
+    Genre("учитель_и_ученик"),
+    Genre("ушастые"),
+    Genre("фантазии"),
+    Genre("фантастика"),
+    Genre("фемдом"),
+    Genre("фестиваль"),
+    Genre("фетиш"),
+    Genre("фистинг"),
+    Genre("фурри"),
+    Genre("футанари"),
+    Genre("футанари_имеет_парня"),
+    Genre("фэнтези"),
+    Genre("хоррор"),
+    Genre("цундере"),
+    Genre("чикан"),
+    Genre("чирлидеры"),
+    Genre("чулки"),
+    Genre("школьная_форма"),
+    Genre("школьники"),
+    Genre("школьницы"),
+    Genre("школьный_купальник"),
+    Genre("щекотка"),
+    Genre("эксгибиционизм"),
+    Genre("эльфы"),
+    Genre("эччи"),
+    Genre("юмор"),
+    Genre("юри"),
+    Genre("яндере"),
+    Genre("яой"),
+)
